fix(#38): 리뷰 피드백 반영
- CompletableFuture 병렬 처리로 성능 개선
- 구체적인 예외 타입(InterruptedException, ExecutionException, TimeoutException) 사용
- price 필드를 BigDecimal로 변경하여 정밀도 보장
- ProductId.of() 정적 팩토리 메서드 사용
- Stream API 사용으로 코드 간결성 향상
- 타임아웃 값을 application.yml 설정으로 분리
- 테스트 코드에서 Map.of()와 List.of() 사용

+++ core/product-core/src/main/java/com/commerce/product/application/usecase/GetProductResponse.java
@@ -5,6 +5,7 @@ import com.commerce.product.domain.model.ProductType;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Value
@@ -22,7 +23,7 @@ public class GetProductResponse {
     public static class ProductOptionResponse {
         String optionId;
         String name;
-        long price;
+        BigDecimal price;
         String currency;
         List<SkuMappingResponse> skuMappings;
         boolean isAvailable;
